deduplicate test helper
